requirements.txt
Flask==2.2.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
requests==2.31.0
python-dotenv==1.0.0
